refactor: break monolithic test_filter_bench into modular compliance tests
- Split 1799-line test_filter_bench.c into focused test modules
- Each test category now in its own file (lifecycle, connection, dataflow, etc.)
- Reduces token usage from ~45k to 1-5k per test file
- Improves maintainability and git history clarity
- Tests continue to pass with same coverage
- Easier to add buffer configuration variations per test
